requirements.txt
# ollama. curl -fsSL https://ollama.ai/install.sh | sh
numpy>=1.26.2
requests>=2.31.0
pypdf>=3.17.4
sentence-transformers>=2.7.0
einops>=0.8.0
chromadb>=0.5.3
gradio==4.44.1
tqdm>=4.66.1
yake>=0.4.8
python-docx==1.1.2
